ST-31-refactor: utils로 QueryKey파일 이동
