+++ src/app/(user)/reservation-status/_components/ReservationStatus.tsx
@@ -2,7 +2,7 @@
 
 import ExistActivity from "@/app/(user)/reservation-status/_components/ExistActivity";
 import NoExistActivity from "@/app/(user)/reservation-status/_components/NoExistActivity";
-import { getMyActivities } from "@/app/(user)/reservation-status/reservationStatus";
+import { getMyActivities } from "@/app/(user)/reservation-status/utils/reservationStatus";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
 function ReservationStatus() {
