+++ src/app/(user)/reservation-status/page.tsx
@@ -1,5 +1,5 @@
 import ReservationStatus from "@/app/(user)/reservation-status/_components/ReservationStatus";
-import { getMyActivities } from "@/app/(user)/reservation-status/reservationStatus";
+import { getMyActivities } from "@/app/(user)/reservation-status/utils/reservationStatus";
 import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query";
 
 async function ReservationStatusPage() {
